warnings fix

+++ Bermuda/Bermuda/StatusTracker.cpp
@@ -54,7 +54,7 @@ void StatusTracker::setAllStats(std::vector<int> stats)
 {
 	if (stats.size() <= this->achievements.size())
 	{
-		for (int i = 0; i < stats.size(); i++)
+		for (size_t i = 0; i < stats.size(); i++)
 		{
 			this->achievements[i]->setAmount(stats[i]);
 		}
